fix: rename events/ to stream-events/ to resolve ES module collision
Renames src/events/ to src/stream-events/ and updates all imports to prevent
tsc-alias from incorrectly transforming Node.js built-in 'events' imports
into relative paths, which was causing CI build failures with ES module
directory import errors.
